# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management API",
        "description": "API for managing translations, locales, and tags",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List translations with filters",
                "operationId": "3c2930eff48d245d2c2747e017f044e6",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Filter by translation key prefix",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Filter by locale code (e.g. en, fr)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Filter by tag name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Number of items per page for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of translations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Translation"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a translation",
                "operationId": "6042ec1977ca7c72336f2258639d0ca4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale_id",
                                    "key",
                                    "value"
                                ],
                                "properties": {
                                    "locale_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "welcome.title"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "Welcome"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": "[1,2]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get a translation",
                "operationId": "7ba5516debb5f8be391ab24c2e35246c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update a translation",
                "operationId": "290e8b744f1aed62ce18a8a711e694a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale_id",
                                    "key",
                                    "value"
                                ],
                                "properties": {
                                    "locale_id": {
                                        "type": "integer"
                                    },
                                    "key": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a translation",
                "operationId": "6be073d8de9bc7a1d9a6859664dd7c16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Translation deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/translations/export": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Export translations for frontend",
                "operationId": "2cfcf3218fd0a9eef9b536e6e164137d",
                "responses": {
                    "200": {
                        "description": "Translations grouped by locale",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Locale": {
                "required": [
                    "id",
                    "code"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "en"
                    },
                    "name": {
                        "type": "string",
                        "example": "English"
                    }
                },
                "type": "object"
            },
            "Tag": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "mobile"
                    }
                },
                "type": "object"
            },
            "Translation": {
                "required": [
                    "id",
                    "key",
                    "value",
                    "locale"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "key": {
                        "type": "string",
                        "example": "welcome.title"
                    },
                    "value": {
                        "type": "string",
                        "example": "Welcome"
                    },
                    "locale": {
                        "$ref": "#/components/schemas/Locale"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "Schemas",
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "This field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}